* class for closest lepton * atempt to replace signature event and global with complete * replace remaining float with double

+++ HeavyHiggs/include/CutVariables.hh
@@ -28,7 +28,7 @@ public:
     int JetNumber() const;
     int BottomNumber() const;
     int Charge() const;
-    float BottomBdt() const;
+    double BottomBdt() const;
     Angle DeltaR() const;
     Mass InvariantMass() const;
     Momentum LeptonPt(int number) const;
