+++ include/plotting/Histograms.hh
@@ -26,19 +26,19 @@ public:
 
     void Initialize(std::string const& path, std::string const& name, bool show_title = false);
 
-    void AddHistogram(std::vector<float> const& values, std::string const& name, Rectangle<float> const& range);
+    void AddHistogram(std::vector<double> const& values, std::string const& name, Rectangle<double> const& range);
 
-    void AddHistogram(std::vector<float> const& values, std::string const& name, Range<float> const& range, bool is_int = false);
+    void AddHistogram(std::vector<double> const& values, std::string const& name, Range<double> const& range, bool is_int = false);
 
     void SetLegend(Orientation orientation, std::string const& title = "");
 
     void Draw();
 
-    void SetXAxis(std::string const& title, Range<float> const& range = Range<float>());
+    void SetXAxis(std::string const& title, Range<double> const& range = Range<double>());
 
-    void SetYAxis(std::string const& title, Range<float> const& range = Range<float>());
+    void SetYAxis(std::string const& title, Range<double> const& range = Range<double>());
 
-    void AddLine(float x_value, std::string const& title = "");
+    void AddLine(double x_value, std::string const& title = "");
 
 private:
 
@@ -56,7 +56,7 @@ private:
 
     std::vector<TLine> lines_;
 
-    Rectangle<float> range_;
+    Rectangle<double> range_;
 
 };
