+++ source/physics/Math.cpp
@@ -13,39 +13,39 @@ namespace boca
 namespace
 {
 
-float Factor(float value, int digits) {
+double Factor(double value, int digits) {
     return std::pow(10., digits - std::ceil(std::log10(std::abs(value))));
 }
 
 }
 
-float FloorToDigits(float value, int digits) {
+double FloorToDigits(double value, int digits) {
     if (value == 0 || value != value) return 0;
-    float factor = Factor(value, digits);
+    double factor = Factor(value, digits);
     return std::floor(value * factor) / factor;
 }
 
-float CeilToDigits(float value, int digits) {
+double CeilToDigits(double value, int digits) {
     if (value == 0 || value != value) return 0;
-    float factor = Factor(value, digits);
+    double factor = Factor(value, digits);
     return std::ceil(value * factor) / factor;
 }
 
-float RoundERROR(float value){
+double RoundERROR(double value){
     return RoundToDigits(value, 2);
 }
 
-float RoundToDigits(float value, int digits) {
+double RoundToDigits(double value, int digits) {
     INFO0;
     if (value == 0 || value != value) return 0;
-    float factor = Factor(value, digits);
+    double factor = Factor(value, digits);
     return std::round(value * factor) / factor;
 }
 
-float RoundToERROR(float value, float error) {
+double RoundToERROR(double value, double error) {
     INFO0;
     if (value == 0 || value != value) return 0;
-    float factor = Factor(error, 2);
+    double factor = Factor(error, 2);
     return std::round(value * factor) / factor;
 }
